refactor(sdk): extract shared genai helpers from maf integration
The gen_ai.* -> ai.* bridge pieces that are framework-neutral (attribute
and event translation tables, message-event synthesis from the
attribute-based GenAI convention, the TranslatedSpan read-only view,
exporter-swapping helpers, and the content-capture env var) move from
the agent_framework package to integrations/genai.py so the Pydantic AI
integration can reuse them instead of growing a second GenAI translator
(the coordination agreed for #2070). agent_framework re-imports
everything under its historical names, so its public API and tests are
unchanged. MAF-specific logic (workflow spans, handoff_to_* heuristics,
span-name mapping) stays put.
